imp(evm): move CallEVM to EVM module
