+++ x/erc20/types/errors.go
@@ -29,9 +29,7 @@ var (
 	ErrUndefinedOwner         = errorsmod.Register(ModuleName, 6, "undefined owner of contract pair")
 	ErrBalanceInvariance      = errorsmod.Register(ModuleName, 7, "post transfer balance invariant failed")
 	ErrUnexpectedEvent        = errorsmod.Register(ModuleName, 8, "unexpected event")
-	ErrABIPack                = errorsmod.Register(ModuleName, 9, "contract ABI pack failed")
-	ErrABIUnpack              = errorsmod.Register(ModuleName, 10, "contract ABI unpack failed")
-	ErrEVMDenom               = errorsmod.Register(ModuleName, 11, "EVM denomination registration")
-	ErrEVMCall                = errorsmod.Register(ModuleName, 12, "EVM call unexpected error")
-	ErrERC20TokenPairDisabled = errorsmod.Register(ModuleName, 13, "erc20 token pair is disabled")
+	ErrEVMDenom               = errorsmod.Register(ModuleName, 9, "EVM denomination registration")
+	ErrEVMCall                = errorsmod.Register(ModuleName, 10, "EVM call unexpected error")
+	ErrERC20TokenPairDisabled = errorsmod.Register(ModuleName, 11, "erc20 token pair is disabled")
 )
