+++ x/evm/keeper/call.go
@@ -1,0 +1,102 @@
+package keeper
+
+import (
+	"encoding/json"
+	"math/big"
+
+	errorsmod "cosmossdk.io/errors"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+
+	"github.com/evmos/evmos/v11/server/config"
+	"github.com/evmos/evmos/v11/x/evm/types"
+)
+
+// CallEVM performs a smart contract method call using given args
+func (k Keeper) CallEVM(
+	ctx sdk.Context,
+	abi abi.ABI,
+	from, contract common.Address,
+	commit bool,
+	method string,
+	args ...interface{},
+) (*types.MsgEthereumTxResponse, error) {
+	data, err := abi.Pack(method, args...)
+	if err != nil {
+		return nil, errorsmod.Wrap(
+			types.ErrABIPack,
+			errorsmod.Wrap(err, "failed to create transaction data").Error(),
+		)
+	}
+
+	resp, err := k.CallEVMWithData(ctx, from, &contract, data, commit)
+	if err != nil {
+		return nil, errorsmod.Wrapf(err, "contract call failed: method '%s', contract '%s'", method, contract)
+	}
+	return resp, nil
+}
+
+// CallEVMWithData performs a smart contract method call using contract data
+func (k Keeper) CallEVMWithData(
+	ctx sdk.Context,
+	from common.Address,
+	contract *common.Address,
+	data []byte,
+	commit bool,
+) (*types.MsgEthereumTxResponse, error) {
+	nonce, err := k.accountKeeper.GetSequence(ctx, from.Bytes())
+	if err != nil {
+		return nil, err
+	}
+
+	gasCap := config.DefaultGasCap
+	if commit {
+		args, err := json.Marshal(types.TransactionArgs{
+			From: &from,
+			To:   contract,
+			Data: (*hexutil.Bytes)(&data),
+		})
+		if err != nil {
+			return nil, errorsmod.Wrapf(errortypes.ErrJSONMarshal, "failed to marshal tx args: %s", err.Error())
+		}
+
+		gasRes, err := k.EstimateGas(sdk.WrapSDKContext(ctx), &types.EthCallRequest{
+			Args:   args,
+			GasCap: config.DefaultGasCap,
+		})
+		if err != nil {
+			return nil, err
+		}
+		gasCap = gasRes.Gas
+	}
+
+	msg := ethtypes.NewMessage(
+		from,
+		contract,
+		nonce,
+		big.NewInt(0), // amount
+		gasCap,        // gasLimit
+		big.NewInt(0), // gasFeeCap
+		big.NewInt(0), // gasTipCap
+		big.NewInt(0), // gasPrice
+		data,
+		ethtypes.AccessList{}, // AccessList
+		!commit,               // isFake
+	)
+
+	res, err := k.ApplyMessage(ctx, msg, types.NewNoOpTracer(), commit)
+	if err != nil {
+		return nil, err
+	}
+
+	if res.Failed() {
+		return nil, errorsmod.Wrap(types.ErrVMExecution, res.VmError)
+	}
+
+	return res, nil
+}
