+++ x/evm/types/errors.go
@@ -42,6 +42,8 @@ const (
 	codeErrGasOverflow
 	codeErrInvalidAccount
 	codeErrInvalidGasLimit
+	codeErrABIPack
+	codeErrABIUnpack
 )
 
 var ErrPostTxProcessing = errors.New("failed to execute post processing")
@@ -91,6 +93,12 @@ var (
 
 	// ErrInvalidGasLimit returns an error if gas limit value is invalid
 	ErrInvalidGasLimit = errorsmod.Register(ModuleName, codeErrInvalidGasLimit, "invalid gas limit")
+
+	// ErrABIPack returns an error if ABI Pack fails
+	ErrABIPack = errorsmod.Register(ModuleName, codeErrABIPack, "contract ABI pack failed")
+
+	// ErrABIUnpack returns an error if ABI Unpack fails
+	ErrABIUnpack = errorsmod.Register(ModuleName, codeErrABIUnpack, "contract ABI unpack failed")
 )
 
 // NewExecErrorWithReason unpacks the revert return bytes and returns a wrapped error
